move classes that interface with api within reverb_api module
